backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.5
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-js-asset==3.0.1
django-ninja==1.3.0
django-ninja-extra==0.22.2
django-ninja-jwt==5.3.5
injector==0.22.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
